Features

+++ src/Controller/TaskController.php
@@ -8,7 +8,8 @@ use App\Service\TaskService;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
+use Symfony\Component\ExpressionLanguage\Expression;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class TaskController extends AbstractController
@@ -21,10 +22,11 @@ class TaskController extends AbstractController
     #[Route('/tasks', name: 'task_list')]
     public function index(): Response
     {
-        $result = $this->taskService->getAll();   
+        $result = $this->taskService->getAll();
         return $this->render('task/list.html.twig', [
             'controller_name' => 'TaskController',
-            'tasks' => $result
+            'tasks' => $result,
+            'user' => $this->getUser(),
         ]);
     }
 
@@ -35,17 +37,18 @@ class TaskController extends AbstractController
         $formTask = $this->createForm(TaskFormType::class, $task);
         $formTask->handleRequest($request);
         if ($formTask->isSubmitted() && $formTask->isValid()) {
-            $this->taskService->saveTask($task);
+            $this->taskService->saveTask($task, $this->getUser());
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
             return $this->redirectToRoute('task_list');
         }
         return $this->render('task/create.html.twig', [
-            'form' => $formTask->createView()
+            'form' => $formTask->createView(),
+            'user' => $this->getUser(),
         ]);
     }
 
     #[Route('/task/{id}/edit', name: 'task_edit')]
-    public function formEdit(#[MapRequestPayload] Task $task = null, int $id, Request $request): Response
+    public function formEdit(Task $task, int $id, Request $request): Response
     {
         $formTask = $this->createForm(TaskFormType::class, $task);
         $formTask->handleRequest($request);
@@ -57,11 +60,12 @@ class TaskController extends AbstractController
         return $this->render('task/edit.html.twig', [
             'form' => $formTask->createView(),
             'task' => $task,
+            'user' => $this->getUser()
         ]);
     }
-    
+
     #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
-    public function toggleTaskAction(Task $task): Response
+    public function toggleTask(Task $task): Response //action
     {
         $this->taskService->toggleTask($task);
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
@@ -69,11 +73,17 @@ class TaskController extends AbstractController
     }
 
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
-    public function deleteTaskAction(Task $task = null): Response
+    #[IsGranted(new Expression('is_granted("ROLE_USER")'))]
+    public function deleteTask(Task $task = null): Response
     {
-        $this->taskService->deleteTask($task);
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        $currentUser = $this->getUser();
+        $taskOwner = $task->getUser();
+        if ($currentUser == $taskOwner or $this->isGranted('ROLE_ADMIN')) {
+            $this->taskService->deleteTask($task);
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        } else {
+            $this->addFlash('error', 'Vous n\'avez pas les droits pour supprimer cette tâche.');
+        }
         return $this->redirectToRoute('task_list');
     }
-
 }
